Rename database.py to create_database.py
This conflicting name was introduced in master
